refactor: Migrate to NetworkService V2 in VerifyIdToken
